+++ src/config/constants/index.ts
@@ -1,0 +1,61 @@
+/**
+ * 配置常量定义
+ * 包含环境变量映射配置和默认值配置
+ */
+
+// 环境变量映射配置
+// 定义哪些变量需要映射到VITE_前缀
+export const ENV_MAPPING = {
+  // 站点配置
+  SITE_TITLE: 'VITE_SITE_TITLE',
+  SITE_DESCRIPTION: 'VITE_SITE_DESCRIPTION',
+  SITE_KEYWORDS: 'VITE_SITE_KEYWORDS',
+  SITE_OG_IMAGE: 'VITE_SITE_OG_IMAGE',
+
+  // GitHub仓库配置
+  GITHUB_REPO_OWNER: 'VITE_GITHUB_REPO_OWNER',
+  GITHUB_REPO_NAME: 'VITE_GITHUB_REPO_NAME',
+  GITHUB_REPO_BRANCH: 'VITE_GITHUB_REPO_BRANCH',
+
+  // 功能配置
+  HOMEPAGE_FILTER_ENABLED: 'VITE_HOMEPAGE_FILTER_ENABLED',
+  HOMEPAGE_ALLOWED_FOLDERS: 'VITE_HOMEPAGE_ALLOWED_FOLDERS',
+  HOMEPAGE_ALLOWED_FILETYPES: 'VITE_HOMEPAGE_ALLOWED_FILETYPES',
+  HIDE_MAIN_FOLDER_DOWNLOAD: 'VITE_HIDE_MAIN_FOLDER_DOWNLOAD',
+  HIDE_DOWNLOAD_FOLDERS: 'VITE_HIDE_DOWNLOAD_FOLDERS',
+
+  // 代理配置
+  DOWNLOAD_PROXY_URL: 'VITE_DOWNLOAD_PROXY_URL',
+  DOWNLOAD_PROXY_URL_BACKUP1: 'VITE_DOWNLOAD_PROXY_URL_BACKUP1',
+  DOWNLOAD_PROXY_URL_BACKUP2: 'VITE_DOWNLOAD_PROXY_URL_BACKUP2',
+
+  // 访问配置
+  USE_TOKEN_MODE: 'VITE_USE_TOKEN_MODE',
+
+  // 开发者配置
+  DEVELOPER_MODE: 'VITE_DEVELOPER_MODE',
+  CONSOLE_LOGGING: 'VITE_CONSOLE_LOGGING'
+} as const;
+
+// 配置默认值
+export const CONFIG_DEFAULTS = {
+  // 站点配置
+  SITE_TITLE: 'Repo-Viewer',
+  SITE_DESCRIPTION: '基于MD3设计语言的GitHub仓库浏览应用',
+  SITE_KEYWORDS: 'GitHub, 仓库, 浏览器, 代码, 查看器',
+  SITE_OG_IMAGE: '/icon.svg',
+
+  // GitHub仓库配置
+  GITHUB_REPO_OWNER: 'UE-DND',
+  GITHUB_REPO_NAME: 'Repo-Viewer',
+  GITHUB_REPO_BRANCH: 'master',
+
+  // 代理配置
+  DOWNLOAD_PROXY_URL: 'https://gh-proxy.com',
+  DOWNLOAD_PROXY_URL_BACKUP1: 'https://ghproxy.com',
+  DOWNLOAD_PROXY_URL_BACKUP2: 'https://raw.staticdn.net',
+
+  // PAT配置
+  PAT_PREFIXES: ['GITHUB_PAT', 'VITE_GITHUB_PAT'] as const,
+  MAX_PAT_NUMBER: 10,
+} as const;
