+++ src/config/core/config-manager.ts
@@ -1,220 +1,17 @@
 /**
- * 配置管理器
- * 统一管理所有应用配置
+ * 配置管理器核心类
+ * 单例模式的配置管理
  */
 
-// 环境变量映射配置
-// 定义哪些变量需要映射到VITE_前缀
-const ENV_MAPPING = {
-  // 站点配置
-  SITE_TITLE: 'VITE_SITE_TITLE',
-  SITE_DESCRIPTION: 'VITE_SITE_DESCRIPTION',
-  SITE_KEYWORDS: 'VITE_SITE_KEYWORDS',
-  SITE_OG_IMAGE: 'VITE_SITE_OG_IMAGE',
-
-  // GitHub仓库配置
-  GITHUB_REPO_OWNER: 'VITE_GITHUB_REPO_OWNER',
-  GITHUB_REPO_NAME: 'VITE_GITHUB_REPO_NAME',
-  GITHUB_REPO_BRANCH: 'VITE_GITHUB_REPO_BRANCH',
-
-  // 功能配置
-  HOMEPAGE_FILTER_ENABLED: 'VITE_HOMEPAGE_FILTER_ENABLED',
-  HOMEPAGE_ALLOWED_FOLDERS: 'VITE_HOMEPAGE_ALLOWED_FOLDERS',
-  HOMEPAGE_ALLOWED_FILETYPES: 'VITE_HOMEPAGE_ALLOWED_FILETYPES',
-  HIDE_MAIN_FOLDER_DOWNLOAD: 'VITE_HIDE_MAIN_FOLDER_DOWNLOAD',
-  HIDE_DOWNLOAD_FOLDERS: 'VITE_HIDE_DOWNLOAD_FOLDERS',
-
-  // 代理配置
-  DOWNLOAD_PROXY_URL: 'VITE_DOWNLOAD_PROXY_URL',
-  DOWNLOAD_PROXY_URL_BACKUP1: 'VITE_DOWNLOAD_PROXY_URL_BACKUP1',
-  DOWNLOAD_PROXY_URL_BACKUP2: 'VITE_DOWNLOAD_PROXY_URL_BACKUP2',
-
-  // 访问配置
-  USE_TOKEN_MODE: 'VITE_USE_TOKEN_MODE',
-
-  // 开发者配置
-  DEVELOPER_MODE: 'VITE_DEVELOPER_MODE',
-  CONSOLE_LOGGING: 'VITE_CONSOLE_LOGGING'
-} as const;
-
-// 配置默认值
-const CONFIG_DEFAULTS = {
-  // 站点配置
-  SITE_TITLE: 'Repo-Viewer',
-  SITE_DESCRIPTION: '基于MD3设计语言的GitHub仓库浏览应用',
-  SITE_KEYWORDS: 'GitHub, 仓库, 浏览器, 代码, 查看器',
-  SITE_OG_IMAGE: '/icon.svg',
-
-  // GitHub仓库配置
-  GITHUB_REPO_OWNER: 'UE-DND',
-  GITHUB_REPO_NAME: 'Repo-Viewer',
-  GITHUB_REPO_BRANCH: 'master',
-
-  // 代理配置
-  DOWNLOAD_PROXY_URL: 'https://gh-proxy.com',
-  DOWNLOAD_PROXY_URL_BACKUP1: 'https://ghproxy.com',
-  DOWNLOAD_PROXY_URL_BACKUP2: 'https://raw.staticdn.net',
-
-  // PAT配置
-  PAT_PREFIXES: ['GITHUB_PAT', 'VITE_GITHUB_PAT'] as const,
-  MAX_PAT_NUMBER: 10,
-} as const;
-
-const runtimeProcessEnv: Record<string, string | undefined> | undefined =
-  typeof process !== 'undefined' && process.env ? process.env : undefined;
-
-const normalizeEnvValue = (value: unknown): string | undefined => {
-  if (typeof value !== 'string') {
-    return undefined;
-  }
-  const trimmed = value.trim();
-  return trimmed.length > 0 ? trimmed : undefined;
-};
-
-// 支持从无前缀变量自动映射到VITE_前缀变量的解析函数
-const resolveEnvWithMapping = (env: Record<string, any>, plainKey: string, fallback: string): string => {
-  // 优先使用VITE_前缀的变量（如果存在）
-  const viteKey = ENV_MAPPING[plainKey as keyof typeof ENV_MAPPING];
-  if (viteKey) {
-    const viteValue = normalizeEnvValue(env[viteKey]);
-    if (viteValue) {
-      return viteValue;
-    }
-
-    // 如果VITE_变量不存在，尝试使用无前缀变量
-    const plainValue = normalizeEnvValue(env[plainKey]);
-    if (plainValue) {
-      return plainValue;
-    }
-
-    // 检查runtime环境变量
-    if (runtimeProcessEnv) {
-      const runtimeViteValue = normalizeEnvValue(runtimeProcessEnv[viteKey]);
-      if (runtimeViteValue) {
-        return runtimeViteValue;
-      }
-
-      const runtimePlainValue = normalizeEnvValue(runtimeProcessEnv[plainKey]);
-      if (runtimePlainValue) {
-        return runtimePlainValue;
-      }
-    }
-  }
-
-  return fallback;
-};
-
-const hasEnvValue = (env: Record<string, any>, keys: string[]): boolean => {
-  for (const key of keys) {
-    if (normalizeEnvValue(env[key])) {
-      return true;
-    }
-  }
-
-  if (runtimeProcessEnv) {
-    for (const key of keys) {
-      if (normalizeEnvValue(runtimeProcessEnv[key])) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-};
-
-
-// 环境变量解析工具
-class EnvParser {
-  static parseBoolean(value: string | undefined): boolean {
-    return value === 'true';
-  }
-
-  static parseStringArray(value: string | undefined): string[] {
-    if (!value) return [];
-    return value.split(',').filter(Boolean).map(item => item.trim());
-  }
-
-  static validateToken(token: any): token is string {
-    return typeof token === 'string' &&
-           token.trim().length > 0 &&
-           token.trim() !== 'your_token_here' &&
-           !token.includes('placeholder');
-  }
-}
-
-// 配置接口定义
-export interface Config {
-  site: {
-    title: string;
-    description: string;
-    keywords: string;
-    ogImage: string;
-  };
-  github: {
-    repoOwner: string;
-    repoName: string;
-    repoBranch: string;
-  };
-  features: {
-    homepageFilter: {
-      enabled: boolean;
-      allowedFolders: string[];
-      allowedFileTypes: string[];
-    };
-    hideDownload: {
-      enabled: boolean;
-      hiddenFolders: string[];
-    };
-  };
-  proxy: {
-    imageProxyUrl: string;
-    imageProxyUrlBackup1: string;
-    imageProxyUrlBackup2: string;
-    healthCheckTimeout: number;
-    validationTimeout: number;
-    healthCheckInterval: number;
-    recoveryTime: number;
-  };
-  access: {
-    useTokenMode: boolean;
-  };
-  developer: {
-    mode: boolean;
-    consoleLogging: boolean;
-  };
-  runtime: {
-    isDev: boolean;
-    isProd: boolean;
-  };
-  tokens: {
-    githubPATs: string[];
-    totalCount: number;
-  };
-}
-
-// 配置变更监听器类型
-export type ConfigChangeListener = (newConfig: Config, oldConfig: Config) => void;
-
-// 调试信息接口
-export interface ConfigDebugInfo {
-  loadedAt: string;
-  environment: 'development' | 'production';
-  configSummary: {
-    siteTitle: string;
-    repoOwner: string;
-    repoName: string;
-    developerMode: boolean;
-    tokenMode: boolean;
-    tokenCount: number;
-  };
-  envVarStatus: Record<string, boolean>;
-  tokenSources: Array<{key: string; hasValue: boolean; isValid: boolean}>;
-}
+import type { Config, ConfigChangeListener, ConfigDebugInfo } from '../types';
+import { CONFIG_DEFAULTS } from '../constants';
+import { EnvParser } from '../utils/env-parser';
+import { resolveEnvWithMapping, hasEnvValue } from '../utils/env-mapping';
 
 /**
  * 配置管理器 - 单例模式
  */
-class ConfigManager {
+export class ConfigManager {
   private static instance: ConfigManager | null = null;
   private config: Config | null = null;
   private listeners: Set<ConfigChangeListener> = new Set();
@@ -427,23 +224,3 @@ class ConfigManager {
 
 // 导出配置管理器实例
 export const configManager = ConfigManager.getInstance();
-
-// 导出工具类
-export { EnvParser };
-
-// 便捷访问函数
-export const getConfig = () => configManager.getConfig();
-export const getSiteConfig = () => getConfig().site;
-export const getGithubConfig = () => getConfig().github;
-export const getFeaturesConfig = () => getConfig().features;
-export const getProxyConfig = () => getConfig().proxy;
-export const getAccessConfig = () => getConfig().access;
-export const getDeveloperConfig = () => getConfig().developer;
-export const getRuntimeConfig = () => getConfig().runtime;
-export const getTokensConfig = () => getConfig().tokens;
-
-// 特殊便捷函数
-export const isDeveloperMode = () => getConfig().developer.mode;
-export const isTokenMode = () => getConfig().access.useTokenMode;
-export const isDevEnvironment = () => getConfig().runtime.isDev;
-export const getGithubPATs = () => getConfig().tokens.githubPATs;
