+++ src/config/index.ts
@@ -3,17 +3,30 @@
  * 提供简洁的配置访问接口
  */
 
-// 导出配置管理器和所有相关内容
+// 导出类型定义
+export type {
+  Config,
+  ConfigChangeListener,
+  ConfigDebugInfo,
+  EnvMappingOptions
+} from './types';
+
+// 导出常量
 export {
-  configManager,
-  EnvParser,
-  
-  // 类型定义
-  type Config,
-  type ConfigChangeListener,
-  type ConfigDebugInfo,
-  
-  // 便捷访问函数
+  ENV_MAPPING,
+  CONFIG_DEFAULTS
+} from './constants';
+
+// 导出工具类和函数
+export { EnvParser } from './utils/env-parser';
+export {
+  applyEnvMappingForVite,
+  resolveEnvWithMapping,
+  hasEnvValue
+} from './utils/env-mapping';
+
+// 导出配置访问函数
+export {
   getConfig,
   getSiteConfig,
   getGithubConfig,
@@ -23,13 +36,14 @@ export {
   getDeveloperConfig,
   getRuntimeConfig,
   getTokensConfig,
-  
-  // 特殊便捷函数
   isDeveloperMode,
   isTokenMode,
   isDevEnvironment,
-  getGithubPATs,
-} from './ConfigManager';
+  getGithubPATs
+} from './utils/config-accessors';
+
+// 导出配置管理器
+export { configManager, ConfigManager } from './core/config-manager';
 
 // 默认导出配置管理器实例
-export { configManager as default } from './ConfigManager';
+export { configManager as default } from './core/config-manager';
