+++ vite.config.ts
@@ -5,90 +5,13 @@ import * as http from 'http'
 import * as https from 'https'
 import { readFileSync } from 'fs'
 // 导入配置管理器
-import { configManager } from './src/config/ConfigManager'
+import { configManager, applyEnvMappingForVite } from './src/config'
 
 type Logger = {
   log: (...args: any[]) => void
   warn: (...args: any[]) => void
   error: (...args: any[]) => void
   info: (...args: any[]) => void
-}
-
-// GitHub PAT 相关配置
-const PAT_BASE_KEY = 'GITHUB_PAT'
-const MAX_PAT_NUMBER = 10
-
-// 映射配置
-const ENV_MAPPING_CONFIG = {
-  // 站点配置
-  SITE_TITLE: 'VITE_SITE_TITLE',
-  SITE_DESCRIPTION: 'VITE_SITE_DESCRIPTION',
-  SITE_KEYWORDS: 'VITE_SITE_KEYWORDS',
-  SITE_OG_IMAGE: 'VITE_SITE_OG_IMAGE',
-
-  // GitHub仓库配置
-  GITHUB_REPO_OWNER: 'VITE_GITHUB_REPO_OWNER',
-  GITHUB_REPO_NAME: 'VITE_GITHUB_REPO_NAME',
-  GITHUB_REPO_BRANCH: 'VITE_GITHUB_REPO_BRANCH',
-
-  // 功能配置
-  HOMEPAGE_FILTER_ENABLED: 'VITE_HOMEPAGE_FILTER_ENABLED',
-  HOMEPAGE_ALLOWED_FOLDERS: 'VITE_HOMEPAGE_ALLOWED_FOLDERS',
-  HOMEPAGE_ALLOWED_FILETYPES: 'VITE_HOMEPAGE_ALLOWED_FILETYPES',
-  HIDE_MAIN_FOLDER_DOWNLOAD: 'VITE_HIDE_MAIN_FOLDER_DOWNLOAD',
-  HIDE_DOWNLOAD_FOLDERS: 'VITE_HIDE_DOWNLOAD_FOLDERS',
-
-  // 代理配置
-  DOWNLOAD_PROXY_URL: 'VITE_DOWNLOAD_PROXY_URL',
-  DOWNLOAD_PROXY_URL_BACKUP1: 'VITE_DOWNLOAD_PROXY_URL_BACKUP1',
-  DOWNLOAD_PROXY_URL_BACKUP2: 'VITE_DOWNLOAD_PROXY_URL_BACKUP2',
-
-  // 访问配置
-  USE_TOKEN_MODE: 'VITE_USE_TOKEN_MODE',
-
-  // 开发者配置
-  DEVELOPER_MODE: 'VITE_DEVELOPER_MODE',
-  CONSOLE_LOGGING: 'VITE_CONSOLE_LOGGING'
-} as const
-
-const normalizeEnvValue = (value?: string): string | undefined => {
-  if (typeof value !== 'string') {
-    return undefined
-  }
-  const trimmed = value.trim()
-  return trimmed.length > 0 ? trimmed : undefined
-}
-
-const syncEnvMapping = (envSource: Record<string, string | undefined>) => {
-  Object.entries(ENV_MAPPING_CONFIG).forEach(([plainKey, viteKey]) => {
-    const plainValue = normalizeEnvValue(envSource[plainKey] ?? process.env[plainKey])
-    const viteValue = normalizeEnvValue(envSource[viteKey] ?? process.env[viteKey])
-
-    if (plainValue && !viteValue) {
-      process.env[viteKey] = plainValue
-      envSource[viteKey] = plainValue
-    }
-  })
-}
-
-// 仅在开发模式下同步 PAT，避免生产构建时将令牌暴露给前端
-const syncPatEnvVariablesForDev = (envSource: Record<string, string | undefined>, isProdLike: boolean) => {
-  if (isProdLike) {
-    return
-  }
-
-  const suffixes = ['', ...Array.from({ length: MAX_PAT_NUMBER }, (_, index) => `${index + 1}`)]
-
-  suffixes.forEach((suffix) => {
-    const plainKey = `${PAT_BASE_KEY}${suffix}`
-    const viteKey = `VITE_${plainKey}`
-    const plainValue = normalizeEnvValue(envSource[plainKey] ?? process.env[plainKey])
-
-    if (plainValue && !process.env[viteKey]) {
-      process.env[viteKey] = plainValue
-      envSource[viteKey] = plainValue
-    }
-  })
 }
 
 const createLogger = (developerMode: boolean): Logger => ({
@@ -153,11 +76,9 @@ function getPackageVersion() {
 export default defineConfig(({ mode }) => {
   const env = loadEnv(mode, process.cwd(), '')
 
-  // 同步变量映射
-  syncEnvMapping(env)
-
+  // 统一由 ConfigManager 执行环境变量映射（SSOT）
   const isProdLike = mode === 'production' || process.env.NODE_ENV === 'production'
-  syncPatEnvVariablesForDev(env, isProdLike);
+  applyEnvMappingForVite(env, { isProdLike })
 
   // 开发者模式配置 - 控制调试信息显示
   // 使用映射后的VITE_前缀变量或原始变量
